Extract zoomable and swipeable dialogs to functions
