+++ ui/src/main/java/com/neaniesoft/vermilion/ui/videos/VideoPlayer.kt
@@ -68,6 +68,6 @@ fun VideoPlayer(video: VideoDescriptor) {
     AndroidView(factory = {
         playerView
     }, update = {
-        playerView.player = player
-    })
+            playerView.player = player
+        })
 }
